feat: update firestore provider piece
This updates it to be compatible with the alpha / settings branches of
klasa as well as ensures that ScheduledTasks are properly written to
firestore as it doesn't support the structure that Klasa provides by
default. That it to say, Firestore exclusively wants JS objects, but by
default the ScheduledTask data has a ScheduledTask object in there which
it doesn't recognize (it even says so, something about when using
"new"). For more details on that read down from
https://discordapp.com/channels/339942739275677727/339948585325953026/617485150221238282 (klasa-support channel)

This was taken from me porting this provider to TS, I removed the TS
typings and restored the original constructor.

I also transformed the functions to async/await as VSCode was suggesting
it anyway and they do read nicer ;)

Lastly I enhanced the documentation, primarily by also including an
alternative method using environment variables for the credentials but
also making it valid JSDoc for VSCode IntelliSense while I was at it.

Signed-off-by: Jeroen Claassens <support@favware.tech>